reworked the switch.py into a subpackage
