+++ .gitignore
@@ -15,4 +15,3 @@ coverage.xml
 
 # Home Assistant configuration
 config/*
-!config/configuration.yaml
