--- custom_components/magic_areas/switch.py
@@ -1,269 +1,0 @@
-"""Platform file for Magic Area's switch entities."""
-
-import logging
-
-from homeassistant.components.climate.const import (
-    ATTR_PRESET_MODE,
-    DOMAIN as CLIMATE_DOMAIN,
-    SERVICE_SET_PRESET_MODE,
-)
-from homeassistant.components.switch import SwitchDeviceClass, SwitchEntity
-from homeassistant.components.switch.const import DOMAIN as SWITCH_DOMAIN
-from homeassistant.config_entries import ConfigEntry
-from homeassistant.const import ATTR_ENTITY_ID, STATE_OFF, STATE_ON, EntityCategory
-from homeassistant.core import HomeAssistant
-from homeassistant.helpers.dispatcher import async_dispatcher_connect
-from homeassistant.helpers.entity_platform import AddEntitiesCallback
-from homeassistant.helpers.event import async_call_later
-
-from custom_components.magic_areas.base.entities import MagicEntity
-from custom_components.magic_areas.base.magic import MagicArea
-from custom_components.magic_areas.const import (
-    CONF_CLIMATE_CONTROL_ENTITY_ID,
-    CONF_CLIMATE_CONTROL_PRESET_CLEAR,
-    CONF_CLIMATE_CONTROL_PRESET_EXTENDED,
-    CONF_CLIMATE_CONTROL_PRESET_OCCUPIED,
-    CONF_CLIMATE_CONTROL_PRESET_SLEEP,
-    CONF_PRESENCE_HOLD_TIMEOUT,
-    DEFAULT_CLIMATE_CONTROL_PRESET_CLEAR,
-    DEFAULT_CLIMATE_CONTROL_PRESET_EXTENDED,
-    DEFAULT_CLIMATE_CONTROL_PRESET_OCCUPIED,
-    DEFAULT_CLIMATE_CONTROL_PRESET_SLEEP,
-    DEFAULT_PRESENCE_HOLD_TIMEOUT,
-    ONE_MINUTE,
-    AreaStates,
-    MagicAreasEvents,
-    MagicAreasFeatureInfoClimateControl,
-    MagicAreasFeatureInfoLightGroups,
-    MagicAreasFeatureInfoMediaPlayerGroups,
-    MagicAreasFeatureInfoPresenceHold,
-    MagicAreasFeatures,
-)
-from custom_components.magic_areas.helpers.area import get_area_from_config_entry
-from custom_components.magic_areas.util import cleanup_removed_entries
-
-_LOGGER = logging.getLogger(__name__)
-
-
-async def async_setup_entry(
-    hass: HomeAssistant,
-    config_entry: ConfigEntry,
-    async_add_entities: AddEntitiesCallback,
-):
-    """Set up the area switch config entry."""
-
-    area: MagicArea | None = get_area_from_config_entry(hass, config_entry)
-    assert area is not None
-
-    switch_entities = []
-
-    if area.has_feature(MagicAreasFeatures.PRESENCE_HOLD):
-        switch_entities.append(PresenceHoldSwitch(area))
-
-    if area.has_feature(MagicAreasFeatures.LIGHT_GROUPS):
-        switch_entities.append(LightControlSwitch(area))
-
-    if area.has_feature(MagicAreasFeatures.MEDIA_PLAYER_GROUPS):
-        switch_entities.append(MediaPlayerControlSwitch(area))
-
-    if area.has_feature(MagicAreasFeatures.CLIMATE_CONTROL):
-        switch_entities.append(ClimateControlSwitch(area))
-
-    if switch_entities:
-        async_add_entities(switch_entities)
-
-    if SWITCH_DOMAIN in area.magic_entities:
-        cleanup_removed_entries(
-            area.hass, switch_entities, area.magic_entities[SWITCH_DOMAIN]
-        )
-
-
-class SwitchBase(MagicEntity, SwitchEntity):
-    """The base class for all the switches."""
-
-    def __init__(self, area: MagicArea) -> None:
-        """Initialize the base switch bits, basic just a mixin for the two types."""
-        MagicEntity.__init__(self, area, domain=SWITCH_DOMAIN)
-        SwitchEntity.__init__(self)
-        self._attr_device_class = SwitchDeviceClass.SWITCH
-        self._attr_should_poll = False
-        self._attr_is_on = False
-
-    async def async_added_to_hass(self) -> None:
-        """Call when entity about to be added to hass."""
-        await super().async_added_to_hass()
-
-        # Restore state
-        last_state = await self.async_get_last_state()
-        if last_state:
-            self._attr_is_on = last_state.state == STATE_ON
-            self._attr_extra_state_attributes = dict(last_state.attributes)
-
-        self.async_write_ha_state()
-        self.schedule_update_ha_state()
-
-    async def async_turn_on(self, **kwargs) -> None:
-        """Turn on presence hold."""
-        self._attr_state = STATE_ON
-        self._attr_is_on = True
-        self.schedule_update_ha_state()
-
-    async def async_turn_off(self, **kwargs) -> None:
-        """Turn off presence hold."""
-        self._attr_state = STATE_OFF
-        self._attr_is_on = False
-        self.schedule_update_ha_state()
-
-
-class ResettableSwitchBase(SwitchBase):
-    """Control the presense/state from being changed for the device."""
-
-    def __init__(self, area: MagicArea) -> None:
-        """Initialize the switch."""
-        super().__init__(area)
-
-        self._timeout_callback = None
-
-        self.async_on_remove(self._clear_timers)
-
-    def _clear_timers(self) -> None:
-        """Remove the timer on entity removal."""
-        if self._timeout_callback:
-            self._timeout_callback()
-
-    async def _timeout_turn_off(self, next_interval):
-        """Turn off the presence hold after the timeout."""
-        if self._attr_state == STATE_ON:
-            await self.async_turn_off()
-
-    async def async_turn_on(self, **kwargs):
-        """Turn on presence hold."""
-        self._attr_state = STATE_ON
-        self._attr_is_on = True
-        self.schedule_update_ha_state()
-
-        timeout = self.area.feature_config(MagicAreasFeatures.PRESENCE_HOLD).get(
-            CONF_PRESENCE_HOLD_TIMEOUT, DEFAULT_PRESENCE_HOLD_TIMEOUT
-        )
-
-        if timeout and not self._timeout_callback:
-            self._timeout_callback = async_call_later(
-                self.hass, timeout * ONE_MINUTE, self._timeout_turn_off
-            )
-
-    async def async_turn_off(self, **kwargs):
-        """Turn off presence hold."""
-        self._attr_state = STATE_OFF
-        self._attr_is_on = False
-        self.schedule_update_ha_state()
-
-        if self._timeout_callback:
-            self._timeout_callback()
-            self._timeout_callback = None
-
-
-class ClimateControlSwitch(SwitchBase):
-    """Switch to enable/disable climate control."""
-
-    feature_info = MagicAreasFeatureInfoClimateControl()
-    _attr_entity_category = EntityCategory.CONFIG
-
-    preset_map: dict[str, str]
-    climate_entity_id: str | None
-
-    def __init__(self, area: MagicArea) -> None:
-        """Initialize the Climate control switch."""
-
-        SwitchBase.__init__(self, area)
-
-        self.climate_entity_id = self.area.feature_config(
-            MagicAreasFeatures.CLIMATE_CONTROL
-        ).get(CONF_CLIMATE_CONTROL_ENTITY_ID, None)
-
-        if not self.climate_entity_id:
-            raise ValueError("Climate entity not set")
-
-        self.preset_map = {
-            AreaStates.CLEAR: self.area.feature_config(
-                MagicAreasFeatures.CLIMATE_CONTROL
-            ).get(
-                CONF_CLIMATE_CONTROL_PRESET_CLEAR, DEFAULT_CLIMATE_CONTROL_PRESET_CLEAR
-            ),
-            AreaStates.OCCUPIED: self.area.feature_config(
-                MagicAreasFeatures.CLIMATE_CONTROL
-            ).get(
-                CONF_CLIMATE_CONTROL_PRESET_OCCUPIED,
-                DEFAULT_CLIMATE_CONTROL_PRESET_OCCUPIED,
-            ),
-            AreaStates.SLEEP: self.area.feature_config(
-                MagicAreasFeatures.CLIMATE_CONTROL
-            ).get(
-                CONF_CLIMATE_CONTROL_PRESET_SLEEP, DEFAULT_CLIMATE_CONTROL_PRESET_SLEEP
-            ),
-            AreaStates.EXTENDED: self.area.feature_config(
-                MagicAreasFeatures.CLIMATE_CONTROL
-            ).get(
-                CONF_CLIMATE_CONTROL_PRESET_EXTENDED,
-                DEFAULT_CLIMATE_CONTROL_PRESET_EXTENDED,
-            ),
-        }
-
-    async def async_added_to_hass(self) -> None:
-        """Call when entity about to be added to hass."""
-        await super().async_added_to_hass()
-
-        self.async_on_remove(
-            async_dispatcher_connect(
-                self.hass, MagicAreasEvents.AREA_STATE_CHANGED, self.area_state_changed
-            )
-        )
-
-    async def area_state_changed(self, area_id, states_tuple):
-        """Handle area state change event."""
-
-        priority_states: list[str] = [
-            AreaStates.SLEEP,
-            AreaStates.EXTENDED,
-            AreaStates.OCCUPIED,
-        ]
-
-        # Handle area clear because the other states doesn't matter
-        if self.area.has_state(AreaStates.CLEAR):
-            return await self.apply_preset(AreaStates.CLEAR)
-
-        # Handle each state top priority to last, returning early
-        for p_state in priority_states:
-            if self.area.has_state(p_state):
-                return await self.apply_preset(p_state)
-
-    async def apply_preset(self, state_name: str):
-        """Set climate entity to given preset."""
-
-        selected_preset: str = self.preset_map[state_name]
-
-        await self.hass.services.async_call(
-            CLIMATE_DOMAIN,
-            SERVICE_SET_PRESET_MODE,
-            {ATTR_ENTITY_ID: self.climate_entity_id, ATTR_PRESET_MODE: selected_preset},
-        )
-
-
-class LightControlSwitch(SwitchBase):
-    """Switch to enable/disable light control."""
-
-    feature_info = MagicAreasFeatureInfoLightGroups()
-    _attr_entity_category = EntityCategory.CONFIG
-
-
-class MediaPlayerControlSwitch(SwitchBase):
-    """Switch to enable/disable media player control."""
-
-    feature_info = MagicAreasFeatureInfoMediaPlayerGroups()
-    _attr_entity_category = EntityCategory.CONFIG
-
-
-class PresenceHoldSwitch(ResettableSwitchBase):
-    """Switch to enable/disable presence hold."""
-
-    feature_info = MagicAreasFeatureInfoPresenceHold()
-    _attr_entity_category = EntityCategory.CONFIG
